+++ src/elements/traits/OrderElementTrait.php
@@ -52,11 +52,11 @@ trait OrderElementTrait
             }
             case 'shippingFullName':
             {
-                return $this->getShippingAddress() ? $this->getShippingAddress()->getFullName() ?? '' : '';
+                return $this->getShippingAddress() ? $this->getShippingAddress()->fullName ?? '' : '';
             }
             case 'billingFullName':
             {
-                return $this->getBillingAddress() ? $this->getBillingAddress()->getFullName() ?? '' : '';
+                return $this->getBillingAddress() ? $this->getBillingAddress()->fullName ?? '' : '';
             }
             case 'shippingBusinessName':
             {